+++ includes/functions-ads.php
@@ -4,34 +4,6 @@
 defined( 'ABSPATH' ) || die;
 
 /**
- * Returns the soon to be removed/deperecated ad units.
- *
- * @access private
- *
- * @since 0.23.0
- *
- * @return array
- */
-function maipub_get_legacy_ad_units() {
-	return [
-		'button'            => 'button',
-		'footer'            => 'footer',
-		'halfpage'          => 'halfpage',
-		'header'            => 'header',
-		'incontent'         => 'incontent',
-		'incontent-wide'    => 'incontent-wide',
-		'infeed'            => 'infeed',
-		'inrecipe'          => 'inrecipe',
-		'medium-rectangle'  => 'medium-rectangle',
-		'micro-bar'         => 'micro-bar',
-		'podcast-footer'    => 'podcast-footer',
-		'podcast-header'    => 'podcast-header',
-		'sponsored-sidebar' => 'sponsored-sidebar',
-		'sidebar'           => 'sidebar',
-	];
-}
-
-/**
  * Renders an ad unit.
  *
  * @access private
@@ -158,6 +130,8 @@ function maipub_get_page_ads_data() {
 						'slug'     => $slug,
 						'location' => $global_location,
 						'content'  => $content,
+						'header'   => get_field( 'maipub_header_scripts' ),
+						'footer'   => get_field( 'maipub_footer_scripts' ),
 					]
 				);
 			}
@@ -169,6 +143,8 @@ function maipub_get_page_ads_data() {
 						'slug'                => $slug,
 						'location'            => $single_location,
 						'content'             => $content,
+						'header'              => get_field( 'maipub_header_scripts' ),
+						'footer'              => get_field( 'maipub_footer_scripts' ),
 						'content_location'    => get_field( 'maipub_single_content_location' ),
 						'content_count'       => get_field( 'maipub_single_content_count' ),
 						'comment_count'       => get_field( 'maipub_single_comment_count' ),
@@ -190,6 +166,8 @@ function maipub_get_page_ads_data() {
 						'slug'          => $slug,
 						'location'      => $archive_location,
 						'content'       => $content,
+						'header'        => get_field( 'maipub_header_scripts' ),
+						'footer'        => get_field( 'maipub_footer_scripts' ),
 						'content_count' => get_field( 'maipub_archive_content_count' ),
 						'content_item'  => get_field( 'maipub_archive_content_item' ),
 						'types'         => get_field( 'maipub_archive_types' ),
@@ -279,6 +257,8 @@ function maipub_validate_ad_conditions_global( $args ) {
 		'slug'     => '',
 		'location' => '',
 		'content'  => '',
+		'header'   => '',
+		'footer'   => '',
 	] );
 
 	// Sanitize.
@@ -287,6 +267,8 @@ function maipub_validate_ad_conditions_global( $args ) {
 		'slug'     => sanitize_key( $args['slug'] ),
 		'location' => esc_html( $args['location'] ),
 		'content'  => trim( maipub_get_processed_ad_content( $args['content'] ) ),
+		'header'   => $args['header'],
+		'footer'   => $args['footer'],
 	];
 
 	return $args;
@@ -313,6 +295,8 @@ function maipub_validate_ad_conditions_single( $args ) {
 		'slug'                => '',
 		'location'            => '',
 		'content'             => '',
+		'header'              => '',
+		'footer'              => '',
 		'content_location'    => 'after',
 		'content_count'       => 6,
 		'content_count'       => 5,
@@ -331,6 +315,8 @@ function maipub_validate_ad_conditions_single( $args ) {
 		'slug'                => sanitize_key( $args['slug'] ),
 		'location'            => esc_html( $args['location'] ),
 		'content'             => trim( $args['content'] ),
+		'header'              => $args['header'],
+		'footer'              => $args['footer'],
 		'content_location'    => esc_html( $args['content_location'] ),
 		'content_count'       => $args['content_count'] ? array_map( 'absint', explode( ',', (string) $args['content_count'] ) ) : [],
 		'comment_count'       => absint( $args['comment_count'] ),
@@ -458,6 +444,8 @@ function maipub_validate_ad_conditions_archive( $args ) {
 		'slug'          => '',
 		'location'      => '',
 		'content'       => '',
+		'header'        => '',
+		'footer'        => '',
 		'content_count' => '',
 		'content_item'  => 'rows',
 		'types'         => [],
@@ -473,6 +461,8 @@ function maipub_validate_ad_conditions_archive( $args ) {
 		'slug'          => sanitize_key( $args['slug'] ),
 		'location'      => esc_html( $args['location'] ),
 		'content'       => trim( $args['content'] ),
+		'header'        => $args['header'],
+		'footer'        => $args['footer'],
 		'content_count' => $args['content_count'] ? array_map( 'absint', explode( ',', (string) $args['content_count'] ) ) : [],
 		'content_item'  => $args['content_item'] ? sanitize_key( $args['content_item'] ) : 'rows',
 		'types'         => $args['types'] ? array_map( 'esc_html', (array) $args['types'] ) : [],
@@ -522,6 +512,13 @@ function maipub_validate_ad_conditions_archive( $args ) {
 
 		// If not already including, check taxonomies if we're restricting to specific taxonomies.
 		if ( ! $include && ! ( $args['taxonomies'] && ( in_array( '*', $args['taxonomies'] ) || in_array( $object->taxonomy, $args['taxonomies'] ) ) ) ) {
+			return [];
+		}
+	}
+	// Author archive.
+	elseif ( is_author() ) {
+		// Bail if not set to show on author archives.
+		if ( ! ( $args['includes'] || in_array( 'author', $args['includes'] ) ) ) {
 			return [];
 		}
 	}